refactor: moved the serve command into seperate file and setup required flag for port
